requirements.txt
apispec==4.7.1
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.9
flask-pf-common==1.0
flask-pf-marshmallow-swagger==1.2
flask-pf-sqlalchemy==1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.2
marshmallow-sqlalchemy==0.26.1
pf-py-object-man==1.1
pf-python-io==1.0
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
